Update archive.sh - remote backup destination dir
The scp command now sends the backup into a corresponding year directory on the server, rather than collecting backups in the evergreen server.

To use this "out-of-the-box," you'll have to ensure that the specific year directory already exists on the remote server. In my case, I just manually created directories up to 2017 using FileZilla. I chose 2017 because our single-server currently runs on Ubuntu 12.04, which has an EOL of 2017 (not that we'll use it up to that point, but that's the latest it should be used, and hopefully by that time I'll have re-visited the backup server at least once :-D).